+++ .storybook/preview-head.html
@@ -1,2 +1,2 @@
 <link rel="stylesheet" href="/themes/default.css" />
-<script type="module" src="/grace.js"></script>
+<script type="module" src="/baklava.js"></script>
