+++ README.md
@@ -1,3 +1,3 @@
-# Grace Next Design System
+# Baklava Design System
 
 Work in progress. Please check [Grace Next Issue](https://github.com/Trendyol/grace/issues/63)
