+++ docs/adding-a-new-icon.stories.mdx
@@ -28,7 +28,7 @@ import prFormImage from './images/pr-form.png';
 
 ## Steps
 
-1. Go to https://github.dev/Trendyol/grace/tree/next In your browser you'll see Visual Studio Code in your browser with `next` branch (our default branch) of Grace.
+1. Go to https://github.dev/Trendyol/grace/tree/next In your browser you'll see Visual Studio Code in your browser with `next` branch (our default branch) of Baklava.
 1. Use three dot menu and `Branch` -> `Create Branch  in Source Control tab
 
 <img src={createBranchImage} />
