+++ docs/commit-rules.stories.mdx
@@ -12,9 +12,9 @@ import { Meta } from '@storybook/addon-docs';
   }}
 />
 
-# Grace Commit Rules
+# Baklava Commit Rules
 
-While developing Grace we use [conventional commit messages](https://www.conventionalcommits.org/en/v1.0.0/). When you commit a change to your local branch, a couple of checks will be done including commit message. If you make a mistake while writing your commit message, you'll see an error message like below and your commit will be cancelled:
+While developing Baklava we use [conventional commit messages](https://www.conventionalcommits.org/en/v1.0.0/). When you commit a change to your local branch, a couple of checks will be done including commit message. If you make a mistake while writing your commit message, you'll see an error message like below and your commit will be cancelled:
 
 ```
 ⧗   input: fix(desing): testing commit
@@ -32,9 +32,9 @@ If you want you can also use prompt cli to build a commit for staged changes. Af
 
 Because of we automatically generate library versions and release notes from commit messages, it's critical to write commit messages in proper way. Please check [@commitlint/config-conventional](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional#subject-full-stop) documentation to see current rules that we use.
 
-## Specific rules for Grace
+## Specific rules for Baklava
 
-We have some special rules defined for Grace.
+We have some special rules defined for Baklava.
 
 ### Scope list
