+++ docs/design-system/colors.stories.mdx
@@ -2,29 +2,29 @@ import { Meta, ColorPalette, ColorItem } from '@storybook/addon-docs';
 
 <Meta title="Design System/Colors" />
 
-# Grace Color Palette
+# Baklava Color Palette
 
-Grace uses a list of defined color with some default values.
+Baklava uses a list of defined color with some default values.
 
 <ColorPalette>
   <ColorItem
-    title="--gr-color-primary"
+    title="--bl-color-primary"
     subtitle="Primary Color"
     colors={{ Primary: '#F27A1A', Hover: '#EF6114' }}
   />
   <ColorItem
-    title="--gr-color-secondary"
+    title="--bl-color-secondary"
     subtitle="Secondary Color"
     colors={{ Secondary: '#273142', Hover: '#0F131A' }}
   />
-  <ColorItem 
-    title="--gr-color-content" 
-    subtitle="Content Colors" 
+  <ColorItem
+    title="--bl-color-content"
+    subtitle="Content Colors"
     colors={{
-        '--primary': '#273142', 
+        '--primary': '#273142',
         '--secondary': '#6E7787',
         '--tertiary': '#95A1B5',
         '--passive': '#AFBBCA',
-    }} 
+    }}
   />
 </ColorPalette>
