+++ docs/design-system/iconography.stories.mdx
@@ -29,7 +29,7 @@ You can check [icon component](/docs/components-icon--single-icon) for usage inf
   {
     icons.map((iconName) => (
       <IconItem key={iconName} name={iconName}>
-        <gr-icon name={iconName}></gr-icon>
+        <bl-icon name={iconName}></bl-icon>
       </IconItem>
     ))
   }
