+++ docs/design-system/typography.stories.mdx
@@ -12,7 +12,7 @@ import { Meta, Typeset } from '@storybook/addon-docs';
   }}
 />
 
-# Grace Typography
+# Baklava Typography
 
 export const typography = {
   type: {
