+++ docs/linting.stories.mdx
@@ -12,7 +12,8 @@ import { Meta } from '@storybook/addon-docs';
   }}
 />
 
-# Grace Testing
+# Baklava testing
+
 > This guide will help contributors about how to lint Grace.
 
 ### Brief
