+++ docs/testing.stories.mdx
@@ -12,8 +12,8 @@ import { Meta } from '@storybook/addon-docs';
   }}
 />
 
-# Grace Testing
-> This guide will help contributors about how to test Grace.
+# Baklava Testing
+> This guide will help contributors about how to test Baklava.
 
 ### Brief
 - Our coverage percentage for tests is `100%`.
