+++ docs/welcome.stories.mdx
@@ -13,6 +13,6 @@ import { Meta, Canvas, Story, Preview } from '@storybook/addon-docs';
   }}
 />
 
-# Welcome to Grace design system
+# Welcome to Baklava Design System
 
 > This document serves as a landing page.
