+++ src/baklava.ts
@@ -1,0 +1,3 @@
+export { default as BlIcon } from './components/icon/bl-icon';
+export { default as BlButton } from './components/button/bl-button';
+export { getIconPath, setIconPath } from './utilities/asset-paths';
