--- src/grace.ts
@@ -1,3 +1,0 @@
-export { default as GrIcon } from './components/icon/gr-icon';
-export { default as GrButton } from './components/gr-button/gr-button';
-export { getIconPath, setIconPath } from './utilities/asset-paths';
